requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.27.0
pydantic-settings>=2.0.0
redis>=5.0.0
httpx>=0.24.0
tenacity>=8.2.0
python-multipart
# Auth
PyJWT>=2.8.0
passlib[bcrypt]
# Testing & Security
pytest>=7.0.0
pytest-asyncio>=0.21.0
greenlet
bandit>=1.7.0
pip-audit>=2.0.0
# Payments & Tasks
stripe>=7.0.0
celery>=5.3.0
flower>=2.0.0

alembic>=1.11.0
